feat: 콘서트 결제 비즈니스 로직 구현

+++ src/main/java/com/sparta/hhplusconcert/infra/point/PaymentRepository.java
@@ -4,6 +4,6 @@ import com.sparta.hhplusconcert.domain.point.entity.PaymentEntity;
 
 public interface PaymentRepository {
 
-  Integer generatePayment(PaymentEntity payment);
+  Long generatePayment(PaymentEntity payment);
 
 }
